```
$ go test -count=10 -run=NONE -bench . -benchmem -cpuprofile=cpu-old.prof | tee old.log
BenchmarkInfo-2                     	 1000000	      1610 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1613 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1609 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1612 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1607 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1610 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1606 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1610 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1607 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1615 ns/op	       0 B/op	       0 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21655 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21622 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21095 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20370 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21872 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21211 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21348 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21635 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	   50000	     21765 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21242 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20877 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     21630 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20162 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     21139 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     21127 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     19864 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20463 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20337 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20896 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     21653 ns/op	   16534 B/op	       5 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7569 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7351 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7318 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7375 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7281 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7559 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7416 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7599 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7356 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7530 ns/op	   16464 B/op	       4 allocs/op
PASS
ok  	github.com/hnakamur/ltsvlog	77.547s
```

```
$ go test -count=10 -run=NONE -bench . -benchmem -cpuprofile=cpu-new.prof | tee new.log
BenchmarkInfo-2                     	 1000000	      1636 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1633 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1629 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1641 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1632 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1631 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1628 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1623 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1623 ns/op	       0 B/op	       0 allocs/op
BenchmarkInfo-2                     	 1000000	      1630 ns/op	       0 B/op	       0 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20746 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21394 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21210 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     21272 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20827 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20140 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20840 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20986 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20946 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     20091 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20144 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20890 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20450 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     19709 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20107 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20116 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20083 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20747 ns/op	   16532 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20817 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithStack-2             	  100000	     20152 ns/op	   16533 B/op	       5 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7143 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7180 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7157 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7310 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7229 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7277 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7226 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7093 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7329 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithUTCTime-2           	  200000	      7090 ns/op	   16464 B/op	       4 allocs/op
PASS
ok  	github.com/hnakamur/ltsvlog	77.284s
```

```
$ benchstat old.log new.log
name                      old time/op    new time/op    delta
Info-2                      1.61µs ± 0%    1.63µs ± 1%  +1.29%  (p=0.000 n=10+10)
ErrWithStackAndUTCTime-2    21.5µs ± 2%    20.8µs ± 4%  -3.02%  (p=0.001 n=9+10)
ErrWithStack-2              20.8µs ± 5%    20.3µs ± 3%  -2.37%  (p=0.029 n=10+10)
ErrWithUTCTime-2            7.44µs ± 2%    7.20µs ± 2%  -3.12%  (p=0.000 n=10+10)

name                      old alloc/op   new alloc/op   delta
Info-2                       0.00B          0.00B         ~     (all equal)
ErrWithStackAndUTCTime-2    16.5kB ± 0%    16.5kB ± 0%  -0.01%  (p=0.001 n=10+10)
ErrWithStack-2              16.5kB ± 0%    16.5kB ± 0%    ~     (p=0.056 n=10+10)
ErrWithUTCTime-2            16.5kB ± 0%    16.5kB ± 0%    ~     (all equal)

name                      old allocs/op  new allocs/op  delta
Info-2                        0.00           0.00         ~     (all equal)
ErrWithStackAndUTCTime-2      5.00 ± 0%      5.00 ± 0%    ~     (all equal)
ErrWithStack-2                5.00 ± 0%      5.00 ± 0%    ~     (all equal)
ErrWithUTCTime-2              4.00 ± 0%      4.00 ± 0%    ~     (all equal)
```